Adjustments at db.h

+++ src/db.h
@@ -264,7 +264,7 @@ public:
         const unsigned char* data = sqlite3_column_text(mStmt, num);
         if (!data)
             return std::string();
-        int size = sqlite3_column_bytes(mStmt, num);
+        size_t size = static_cast<size_t>(sqlite3_column_bytes(mStmt, num));
         return std::string((const char*)data, size);
     }
     bool hasBlobCol(int num)
@@ -302,7 +302,7 @@ public:
         const void* data = sqlite3_column_blob(mStmt, num);
         if (!data)
             return 0;
-        size_t size = sqlite3_column_bytes(mStmt, num);
+        size_t size = static_cast<size_t>(sqlite3_column_bytes(mStmt, num));
         if (size > buflen)
             throw std::runtime_error("blobCol: Insufficient buffer space for blob: required "+
                 std::to_string(size)+", provided "+std::to_string(buflen));
@@ -310,8 +310,9 @@ public:
         return size;
     }
 
-    uint64_t uint64Col(int num) { return (uint64_t)sqlite3_column_int64(mStmt, num);}
-    unsigned int uintCol(int num) { return (unsigned int)sqlite3_column_int(mStmt, num);}
+    // TODO: ensure that callers invoke the right prototype to avoid unnecessary castings
+    uint64_t uint64Col(int num) { return static_cast<uint64_t>(sqlite3_column_int64(mStmt, num));}
+    unsigned int uintCol(int num) { return static_cast<unsigned int>(sqlite3_column_int(mStmt, num));}
     bool isNullColumn (int num) const { return sqlite3_column_type(mStmt, num) == SQLITE_NULL; }
     int getColumnBytes (int num) const { return sqlite3_column_bytes(mStmt, num); }
 };
